Refactor Pet and Profile Tabs to separate components
