+++ client/src/App.js
@@ -3,7 +3,7 @@ import TopBar from './components/toolbar/TopBar.js'
 import Login from './components/signin/Login.js'
 import Signup from './components/signin/Signup.js'
 import ResetPassword from './components/signin/ResetPassword.js'
-import ProfilePage from './components/pages/ProfilePage.js'
+import ProfilePage from './components/profile/ProfilePage.js'
 import Playgroups from './components/pages/Playgroups.js'
 import Shop from './components/pages/Shop.js'
 import SettingsPage from './components/settings/SettingsPage.js'
